modified sender and common 429 to use real world arinc encoding and data.

+++ sender/flight_state.h
@@ -4,7 +4,7 @@
 extern int throttle_pct;
 extern int bank_angle;
 extern int pitch_angle;
-extern int altitude;
+extern double altitude;
 extern int rpm;
 
 #endif // FLIGHT_STATE_H
